Gif data class removed, using the MutableGifMediaViewModel instead. Added more documentation
