+++ WordPress/src/main/java/org/wordpress/android/viewmodel/gifs/GifPickerDataSource.kt
@@ -1,9 +1,9 @@
-package org.wordpress.android.viewmodel.gif
+package org.wordpress.android.viewmodel.gifs
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.paging.PositionalDataSource
-import org.wordpress.android.viewmodel.gif.provider.GifProvider
+import org.wordpress.android.viewmodel.gifs.provider.GifProvider
 
 /**
  * The PagedListDataSource that is created and managed by [GifPickerDataSourceFactory]
@@ -67,8 +67,8 @@ class GifPickerDataSource(
         initialLoadError = null
         _rangeLoadErrorEvent.postValue(null)
 
-        // Do not do any API call if the [searchQuery] is empty
         when {
+            // Do not do any API call if the [searchQuery] is empty
             searchQuery.isBlank() ->
                 callback.onResult(emptyList(), startPosition, 0)
