+++ WordPress/src/main/java/org/wordpress/android/viewmodel/gifs/GifPickerViewModel.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.viewmodel.gif
+package org.wordpress.android.viewmodel.gifs
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
