+++ WordPress/src/test/java/org/wordpress/android/viewmodel/gifs/GifPickerViewModelTest.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.viewmodel.gif
+package org.wordpress.android.viewmodel.gifs
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.paging.PositionalDataSource.LoadInitialCallback
@@ -19,8 +19,8 @@ import org.mockito.Mock
 import org.mockito.junit.MockitoJUnitRunner
 import org.wordpress.android.fluxc.model.MediaModel
 import org.wordpress.android.util.NetworkUtilsWrapper
-import org.wordpress.android.viewmodel.gif.GifPickerViewModel.EmptyDisplayMode
-import org.wordpress.android.viewmodel.gif.GifPickerViewModel.State
+import org.wordpress.android.viewmodel.gifs.GifPickerViewModel.EmptyDisplayMode
+import org.wordpress.android.viewmodel.gifs.GifPickerViewModel.State
 import java.util.Random
 import java.util.UUID
